% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{as.owin}
\alias{as.owin}
\title{accepts an extent object and returns it as an owin object used
in the spatstat package.}
\usage{
as.owin(x, extentMultiplier = 1.1)
}
\arguments{
\item{x}{original extent object}

\item{extentMultiplier}{optional extent multiplier to expand our owin beyond the extent of x}
}
\description{
accepts an extent object and returns it as an owin object used
in the spatstat package.
}

