% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{cropRasterByPolygons}
\alias{cropRasterByPolygons}
\title{Accepts a raster and SpatialPolygonDataFrame object, iterates over each polygon feature, creating rasters
for each step.  A raster list is returned to the user. Useful for parsing out climate/elevation data, county-by-county,
for an entire state and then processing with the parallel package.}
\usage{
cropRasterByPolygons(r = NULL, s = NULL, field = NULL, write = F,
  parallel = F)
}
\description{
Accepts a raster and SpatialPolygonDataFrame object, iterates over each polygon feature, creating rasters
for each step.  A raster list is returned to the user. Useful for parsing out climate/elevation data, county-by-county,
for an entire state and then processing with the parallel package.
}

