% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{extractDensities}
\alias{extractDensities}
\title{extract quantiles from a continuous raster surface as spatial polygons using GDAL}
\usage{
extractDensities(x, s = 5, d = 15, p = c(0.5, 0.9))
}
\description{
extract quantiles from a continuous raster surface as spatial polygons using GDAL
}

