% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{getGDALtoolByName}
\alias{getGDALtoolByName}
\title{built-in (hidden) function that will accept a GDAL tool by name and attempt to find it within the user's current PATH}
\usage{
getGDALtoolByName(x = NULL)
}
\arguments{
\item{x}{GDAL tool name (e.g., gdal_proximity.py)}
}
\description{
built-in (hidden) function that will accept a GDAL tool by name and attempt to find it within the user's current PATH
}

