% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{getPythonPath}
\alias{getPythonPath}
\title{built-in (hidden) function that will parse the system path for all occurrences of python and take the first occurrence as the version to use}
\usage{
getPythonPath()
}
\description{
built-in (hidden) function that will parse the system path for all occurrences of python and take the first occurrence as the version to use
}

