% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{include}
\alias{include}
\title{built-in (hidden) wrapper function for require that will rudely attempt to install missing packages.}
\usage{
include(x, from = "cran", repo = NULL)
}
\arguments{
\item{x}{specifies the package name}

\item{from}{specifies whether we are fetching from CRAN or GitHub}

\item{repo}{specifies the repository URL to use.  Will pick a few good ones by default.}
}
\description{
built-in (hidden) wrapper function for require that will rudely attempt to install missing packages.
}

