
\name{landscapeAnalysis-package}
\alias{landscapeAnalysis-package}
\alias{landscapeAnalysis}

\docType{package}

\title{ Landscape Analysis 
}
\description{
\pkg{landscape Analysis} Various tools in R/Python that are useful for getting stuff done in landscape ecology.

}
\details{
\tabular{ll}{
Package: \tab landscapeAnalysis\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-22\cr
License: \tab GPL-3 \cr
}

}
\author{
Kyle Taylor 
<kyle.taylor@pljv.org>
}

\keyword{ package }
