% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeMetrics.R
\name{lExtract}
\alias{lExtract}
\title{takes spatial features (e.g., SpatialPoints) from y= and extracts across a list of raster scenes specified by X=
handy for extracting data across the spatial extent of rasters without having to mosaic and create a 'stack' object.}
\usage{
lExtract(X = NULL, y = NULL, fun = mean)
}
\arguments{
\item{X}{list of rasters}

\item{y}{Spatial* object to extract across the 'X' raster series.}
}
\description{
takes spatial features (e.g., SpatialPoints) from y= and extracts across a list of raster scenes specified by X=
handy for extracting data across the spatial extent of rasters without having to mosaic and create a 'stack' object.
}

