% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeMetrics.R
\name{lReclass}
\alias{lReclass}
\title{accepts a list of rasters and reclasses each raster to a binary surface where 1=focal cover value and
0=everything else}
\usage{
lReclass(x = NULL, inValues = NULL, nomatch = NA)
}
\description{
accepts a list of rasters and reclasses each raster to a binary surface where 1=focal cover value and
0=everything else
}

