% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeMetrics.R
\name{metricsListToVector}
\alias{metricsListToVector}
\title{Internal (hidden) function that parses the list output from lCalculateLandscapeMetrics() by metric name, returning the results as a vector.}
\usage{
metricsListToVector(x, metric = "total.area", class = 1)
}
\description{
Internal (hidden) function that parses the list output from lCalculateLandscapeMetrics() by metric name, returning the results as a vector.
}

