% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{multiplyExtent}
\alias{multiplyExtent}
\title{an extent multiplier that will expand an extent object by a user-specified multiple}
\usage{
multiplyExtent(x, extentMultiplier = 1.1)
}
\arguments{
\item{extentMultiplier}{multiple (anything greater than 1 will extend by 1-multiple percent)}
}
\description{
an extent multiplier that will expand an extent object by a user-specified multiple
}

