% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{rasterToPolygons}
\alias{rasterToPolygons}
\title{convert a raster to polygons using either 'R' or 'GDAL'.  GDAL is the (faster) default selection}
\usage{
rasterToPolygons(r = NULL, method = "gdal")
}
\arguments{
\item{r}{a raster object that we will convert to a polygon}
}
\description{
convert a raster to polygons using either 'R' or 'GDAL'.  GDAL is the (faster) default selection
}

