% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeMetrics.R
\name{spatialPolygonsToGrid}
\alias{spatialPolygonsToGrid}
\title{A quick wrapper funtion for HexPoints2SpatialPolygons(), because I can never remember the name of
the HexPoints2SpatialPolygons function when I need it.}
\usage{
spatialPolygonsToGrid(s = NULL, n = NULL, area = NULL,
  type = "hexagonal")
}
\description{
A quick wrapper funtion for HexPoints2SpatialPolygons(), because I can never remember the name of
the HexPoints2SpatialPolygons function when I need it.
}

