% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalSpatialAnalysis.R
\name{splitExtent}
\alias{splitExtent}
\title{When working with SDB queries for large areas, we consistently lose a lot of data... particularly for large
counties.  This gets around that by splitting an extent object into adjacent quarters so that we can download and
merge our raster segments later.  For small counties, this is inefficient.  But its better than just flatly attempting downloads}
\usage{
splitExtent(e = NULL, multiple = 2)
}
\description{
When working with SDB queries for large areas, we consistently lose a lot of data... particularly for large
counties.  This gets around that by splitting an extent object into adjacent quarters so that we can download and
merge our raster segments later.  For small counties, this is inefficient.  But its better than just flatly attempting downloads
}

