% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeMetrics.R
\name{subsampleSurface}
\alias{subsampleSurface}
\title{Accepts a RasterLayer (and optional input pts=SpatialPoints object) and either generates a
buffered point pattern process across its surface or buffers the pts= data and performs an extraction
across the raster surface using those data.  Returns a list containing the extractions.}
\usage{
subsampleSurface(x = NULL, pts = NULL, n = 100, type = "random",
  width = NULL, DEBUG = F)
}
\description{
Accepts a RasterLayer (and optional input pts=SpatialPoints object) and either generates a
buffered point pattern process across its surface or buffers the pts= data and performs an extraction
across the raster surface using those data.  Returns a list containing the extractions.
}

