% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapeMetrics.R
\name{lExtract}
\alias{lExtract}
\title{takes spatial features (e.g., SpatialPoints) from y= and extracts across a list of raster scenes specified by X=
handy for extracting data across the spatial extent of scenes without having to mosaic the rasters into a single surface.}
\usage{
lExtract(X = NULL, y = NULL, fun = mean)
}
\description{
takes spatial features (e.g., SpatialPoints) from y= and extracts across a list of raster scenes specified by X=
handy for extracting data across the spatial extent of scenes without having to mosaic the rasters into a single surface.
}

